+++ app/batid/management/commands/import_bdtopo_dpt.py
@@ -15,22 +15,3 @@ class Command(BaseCommand):
     def handle(self, *args, **options):
         tasks = create_tasks_list(options["dpt"])
         chain(*tasks)()
-
-
-def create_tasks_list(dpt, bulk_launch_uuid=None):
-
-    most_recent_date = bdtopo_release_before()
-    src_params = bdtopo_src_params(dpt, most_recent_date)
-
-    tasks = []
-    tasks.append(
-        Signature("batid.tasks.dl_source", args=["bdtopo", src_params], immutable=True)
-    )
-    tasks.append(
-        Signature(
-            "batid.tasks.import_bdtopo",
-            args=[src_params, bulk_launch_uuid],
-            immutable=True,
-        )
-    )
-    return tasks
