+++ app/batid/services/source.py
@@ -236,92 +236,3 @@ class BufferToCopy(Source):
         raise Exception(
             f"Can't write buffer, data rows must be a list or a dict, {type(data)} given"
         )
-
-
-def bdtopo_src_params(dpt: str, date: str) -> dict:
-
-    dpt = dpt.zfill(3)
-    projection = _bdtopo_dpt_projection(dpt)
-
-    return {
-        "dpt": dpt,
-        "projection": projection,
-        "date": date,
-    }
-
-
-def _bdtopo_dpt_projection(dpt: str) -> str:
-
-    default_proj = "LAMB93"
-
-    projs = {
-        "971": "RGAF09UTM20",
-        "972": "RGAF09UTM20",
-        "973": "UTM22RGFG95",
-        "974": "RGR92UTM40S",
-        "975": "RGSPM06U21",
-        "976": "RGM04UTM38S",
-        "977": "RGAF09UTM20",
-        "978": "RGAF09UTM20",
-    }
-
-    return projs.get(dpt, default_proj)
-
-
-def bdtopo_release_before(before: date = None) -> str:
-
-    # If no date is provided, we use the current date
-    if before is None:
-        before = datetime.now().date()
-
-    # First we get an ordred list of all release dates
-    release_dates = sorted(
-        [datetime.strptime(date, "%Y-%m-%d").date() for date in _bdtopo_release_dates()]
-    )
-
-    for idx, date in enumerate(release_dates):
-        if date >= before:
-
-            # Return the previous release date
-            return release_dates[idx - 1].strftime("%Y-%m-%d")
-
-
-def _bdtopo_release_dates() -> list:
-
-    return [
-        #
-        "2024-03-15",
-        "2024-06-15",
-        "2024-09-15",
-        "2024-12-15",
-        #
-        "2025-03-15",
-        "2025-06-15",
-        "2025-09-15",
-        "2025-12-15",
-        #
-        "2026-03-15",
-        "2026-06-15",
-        "2026-09-15",
-        "2026-12-15",
-        #
-        "2027-03-15",
-        "2027-06-15",
-        "2027-09-15",
-        "2027-12-15",
-        #
-        "2028-03-15",
-        "2028-06-15",
-        "2028-09-15",
-        "2028-12-15",
-        #
-        "2029-03-15",
-        "2029-06-15",
-        "2029-09-15",
-        "2029-12-15",
-        #
-        "2030-03-15",
-        "2030-06-15",
-        "2030-09-15",
-        "2030-12-15",
-    ]
